backend/requirements.txt
flask
flask-cors
chromadb
openai
python-dotenv
requests
datetime
uuid
PyPDF2
gspread
google-auth
google-api-python-client
google-auth-oauthlib
pymongo
dropbox
soundfile
flask-socketio
langchain
langchain-openai